Validate OpenAPI specification in CI
Fixes #41

When running CI, after linting the code, validate the API specification. Ensure it is valid per the OpenAPI Specification/OAS (semantic check).

Note that by validating OAS, we are also ensuring it is valid YAML (syntactic check). We do not need a separate tool or check for this.

This step is important since the API application uses the committee gem to parse the spec and enforce specified constraints, e.g., controller parameters, and the committee gem does not provide the most useful feedback when the spec is not valid.