remove test of built-in date math methods
noticed that test was flappy (Date.today returns UTC, but days_ago.days.ago.to_date returns local time zone, so result is off late in the evening in pacific time).  went to fix it and then noticed it also seemed to be testing built in date conversion and math functionality, so removed that expectation.