Skip managing files when Cocina file URIs present
Connects to sul-dlss/sdr-client#184

Currently, sdr-api assumes it needs to be a file-managing proxy for every request that comes through. That does not work if a user is using sdr-client as a CLI and, e.g., only wants to change the APO for an item. This commit adds some smarts to sdr-api allowing it to know whether or not to manage files by proxy or not.