Use different database names per Rails env
Else db:setup will fail because it tries to create both the dev and test databases... which used to be the same. In so doing, keep the production database name the same. But it does not matter much since we override this file in puppet on deployments. The config/database.yml changes are now more consistent with our other codebases, e.g., DSA.